backend/requirements-web.txt
# Core Web Dependencies
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# HTTP Client (for RAG API)
httpx==0.28.1

# Session Management (Optional)
redis==5.2.1

# Security
slowapi==0.1.9  # Rate limiting
python-decouple==3.8

# Utilities
python-dotenv==1.0.1

# Logging
loguru==0.7.3